create services

+++ src/main/java/com/example/springdata/restaurantbackend/Repository/NotificationRepository.java
@@ -3,5 +3,5 @@ package com.example.springdata.restaurantbackend.Repository;
 import com.example.springdata.restaurantbackend.Entity.Notification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NotificationRepository extends JpaRepository<Notification, Integer> {
+public interface NotificationRepository extends JpaRepository<Notification, Long> {
 }
